backend/requirements.txt
fastapi
uvicorn[standard]

numpy
pillow
opencv-python-headless

torch
torchvision
ultralytics

scikit-learn
faiss-cpu

python-multipart
requests
tqdm

open_clip_torch
git+https://github.com/openai/CLIP.git